Move seperate_fields function to molo.utils
TODO: remove dependency on api Constants in function
